+++ src/main/java/open/commons/spring/web/beans/authority/IAuthorizedDataHandler.java
@@ -18,7 +18,7 @@
  *
  * This file is generated under this project, "open-commons-spring-web".
  *
- * Date  : 2025. 5. 20. 오후 4:41:27
+ * Date  : 2025. 9. 18. 오후 12:46:42
  *
  * Author: parkjunhong77@gmail.com
  * 
@@ -26,38 +26,40 @@
 
 package open.commons.spring.web.beans.authority;
 
+import open.commons.spring.web.authority.AuthorizedData;
 import open.commons.spring.web.authority.AuthorizedField;
 
 /**
- * {@link AuthorizedField#fieldHandleBean()}에서 사용되는 인터페이스.
+ * {@link AuthorizedData#handleBean()}에서 사용되는 인터페이스.
  * 
- * @since 2025. 5. 20.
+ * @since 2025. 9. 18.
  * @version 0.8.0
  * @author parkjunhong77@gmail.com
  */
-public interface IUnauthorizedFieldHandler {
+public interface IAuthorizedDataHandler {
 
     /**
-     * 데이터 처리방식에 따라 데이터를 처리한 결과를 제공합니다.
+     * 데이터 처리 방식에 따라서 {@link AuthorizedField#fieldHandleBean()}에 의해서 처리된 데이터를 원복하여 제공합니다. <br>
      * 
      * <pre>
      * [개정이력]
      *      날짜    	| 작성자	|	내용
      * ------------------------------------------
-     * 2025. 5. 26.		박준홍			최초 작성
+     * 2025. 9. 18.		박준홍			최초 작성
      * </pre>
      *
      * @param handle
-     *            데이터 처리방식 식별정보
-     * @param data
-     *            데이터
+     *            데이터 처리 방식 식별정보
+     * @param value
+     *            {@link AuthorizedField#fieldHandleBean()}에 의해서 처리된 데이터
      * @return
      * @throws UnsupportedOperationException
      *             지원하지 않는 <code>handle</code>을 전달받았을 때.
-     * @since 2025. 5. 26.
+     *
+     * @since 2025. 9. 18.
      * @version 0.8.0
      * @author Park, Jun-Hong parkjunhong77@gmail.com
      */
-    Object handleObject(int handle, Object data) throws UnsupportedOperationException;
+    Object restoreValue(int handle, Object value) throws UnsupportedOperationException;
 
 }