+++ src/main/java/open/commons/spring/web/config/CustomWebMvcAutoConfiguration.java
@@ -26,11 +26,31 @@
 
 package open.commons.spring.web.config;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+
+import javax.annotation.Nonnull;
+import javax.validation.constraints.NotNull;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.core.env.Environment;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerAdapter;
+
+import open.commons.core.utils.MapUtils;
+import open.commons.spring.web.beans.resolver.AuthorizedDataArgumentResolver;
+import open.commons.spring.web.beans.resolver.AuthorizedDataModelAttributeResolver;
+import open.commons.spring.web.beans.resolver.IAuthorizedDataResolver;
 
 /**
  * 
@@ -42,6 +62,10 @@ import org.springframework.core.env.Environment;
 // 하위 클래스가 @Configuration/@Bean으로 등록돼 있으면 이 빈은 건너뜀
 @ConditionalOnMissingBean(CustomWebMvcConfigurer.class)
 public class CustomWebMvcAutoConfiguration {
+
+    public static final String BEAN_QUALIFIER_AUTHORIZED_DATA_RESOLVERS = "open.commons.spring.web.config.CustomWebMvcAutoConfiguration#AUTHORIZED_DATA_RESOLVERS";
+
+    private static final Logger logger = LoggerFactory.getLogger(CustomWebMvcAutoConfiguration.class);
 
     private final ApplicationContext context;
     private final Environment environment;
@@ -70,7 +94,63 @@ public class CustomWebMvcAutoConfiguration {
     }
 
     @Bean
-    public CustomWebMvcConfigurer customWebMvcConfigurer() {
-        return new CustomWebMvcConfigurer(this.context, this.environment);
+    @Primary
+    AuthorizedDataArgumentResolver authorizedDataArgumentRevolser(ApplicationContext context) {
+        AuthorizedDataArgumentResolver resolver = new AuthorizedDataArgumentResolver(context);
+        logger.info("[authorized-resources] authorized-data-argument-resolver={}", resolver);
+        return resolver;
     }
+
+    @Bean
+    @Primary
+    AuthorizedDataModelAttributeResolver authorizedDataModelAttributeResolver(ApplicationContext context) {
+        AuthorizedDataModelAttributeResolver resolver = new AuthorizedDataModelAttributeResolver(context);
+        logger.info("[authorized-resources] authorized-data-model-attribute-resolver={}", resolver);
+        return resolver;
+    }
+
+    @Bean(BEAN_QUALIFIER_AUTHORIZED_DATA_RESOLVERS)
+    @Primary
+    List<IAuthorizedDataResolver> authorizedDataResolver(@NotNull @Nonnull Map<String, IAuthorizedDataResolver> single //
+            , @NotNull @Nonnull Map<String, List<IAuthorizedDataResolver>> multi) {
+        List<IAuthorizedDataResolver> resolvers = MapUtils.toList(single, multi);
+        logger.info("[authorized-resources] authorized-data-resolvers={}", resolvers);
+        return resolvers;
+    }
+
+    @Bean
+    public CustomWebMvcConfigurer customWebMvcConfigurer() {
+        CustomWebMvcConfigurer c = new CustomWebMvcConfigurer(this.context, this.environment);
+        logger.info("[web-mvc-configurer] custom-web-mvc-configurer={}", c);
+        return c;
+    }
+
+    @Bean
+    public BeanPostProcessor reorderArgumentResolvers() {
+        return new BeanPostProcessor() {
+            @Override
+            public Object postProcessAfterInitialization(Object bean, String beanName) {
+                if (bean instanceof RequestMappingHandlerAdapter) {
+                    RequestMappingHandlerAdapter adapter = (RequestMappingHandlerAdapter) bean;
+                    List<HandlerMethodArgumentResolver> resolvers = new ArrayList<>(adapter.getArgumentResolvers());
+                    Collections.sort(resolvers, new Comparator<HandlerMethodArgumentResolver>() {
+                        @Override
+                        public int compare(HandlerMethodArgumentResolver o1, HandlerMethodArgumentResolver o2) {
+                            if (o1 instanceof IAuthorizedDataResolver) {
+                                return -1;
+                            } else if (o2 instanceof IAuthorizedDataResolver) {
+                                return 1;
+                            } else {
+                                return 0;
+                            }
+                        }
+                    });
+
+                    adapter.setArgumentResolvers(resolvers);
+                }
+                return bean;
+            }
+        };
+    }
+
 }